DUAP: continuous integration, and three checks that had never been run
.github/workflows/duap.yml adds seven jobs: Rust build/fmt/clippy/tests/
docs, conformance vectors against the generator, the independent Go
implementation verifying the committed vectors, five TLA+ configurations,
generated-file idempotence, the documentation and personal-data rules, and
cargo-deny.

Writing it was worth more than having it. Several of these checks already
existed and had only ever inspected staged diffs; running them over the
whole tree for the first time found a defect in every one.

  CI-01  Generated code was not formatter-idempotent, so rule 06's
         staleness check could not have worked. gen_code.py emitted
         unformatted source, cargo fmt reformatted taxonomy.rs in place,
         and the next regeneration undid it -- a 161-line phantom diff on
         every formatting-adjacent run, which is the false alarm that
         teaches people to ignore a check. The Go emitter was worse: the
         committed taxonomy.go had never been gofmt-clean. The generator
         now runs rustfmt and gofmt over its own output, so generated code
         is a fixed point, and CI diffs it.

  The personal-data hook's payment-card pattern used \b, which matches
  after a decimal point, so the float 2.5560241107087633 in a research
  results file read as a Mastercard number. Fixed with lookarounds that
  exclude digit runs inside longer numbers.

  The maturity-marker hook's vocabulary omitted PROTOTYPE, which rule 10
  has and rule 01 did not. The VS-8 status commit had written PROTOTYPE
  into four crate headers, so those four read as having no marker at all
  and the check silently passed. ADR-0015 aligns the two vocabularies and
  explains why the distinction the other option defended was not paying
  for itself.

Both hooks gained an --all mode that scans every tracked file and fails
rather than warns. Advisory is right at commit time, where a false block
trains people to bypass hooks; in CI the file set is stable and a missing
marker is a defect.

Seven modules had no maturity marker at all -- error.rs and main.rs files
the diff-scoped hook had never seen. Marked.

Also in this commit, to make the gates pass:

  clippy across the workspace at -D warnings. Most were mechanical. Two
  were judgement: NegotiationMessage's large variant is left unboxed with
  the reason written down, and SubjectAgent::authorize_with's eight
  positional arguments became a GrantPolicy struct, because an expiry and
  a revocation policy are easy to transpose and a transposed grant is one
  that expires when it should have been revocable.

  deny.toml: permissive licences only, advisories and yanked crates deny,
  wildcards deny, unknown sources deny. Enforcing it surfaced that seven
  intra-workspace path dependencies had no version, which makes them
  unpublishable; all now carry one.

264 tests pass, clippy clean, fmt clean, docs clean, 93 vectors, 77/77 in
the Go verifier, five model-check configurations behaving as required.

+++ duap/crates/duap-model/src/pricing.rs
@@ -212,7 +212,7 @@ impl PricingRule {
                         reason: "a revenue share must not be negative".into(),
                     });
                 }
-                if share.num as i128 > share.den as i128 {
+                if share.num > share.den as i128 {
                     return Err(ModelError::Invalid {
                         field: "pricing.share",
                         reason: "a revenue share above 100% is rejected as a modelling error"
